Share telemetry helpers
