+++ packages/app/src/telemetry/otlp.ts
@@ -1,9 +1,2 @@
-export type OtlpSignal = "logs" | "metrics" | "traces";
-
-export function normalizeOtlpOrigin(origin: string): string {
-  return origin.replace(/\/+$/, "");
-}
-
-export function buildOtlpSignalUrl(origin: string, signal: OtlpSignal): string {
-  return `${normalizeOtlpOrigin(origin)}/v1/${signal}`;
-}
+export type { OtlpSignal } from "./shared.ts";
+export { buildOtlpSignalUrl, normalizeOtlpOrigin } from "./shared.ts";
