+++ packages/docs/content/docs/cli/commands.mdx
@@ -42,10 +42,6 @@ description: Reference for all Everr CLI commands.
   - `--tail <n>` how many lines from the bottom
   - `--limit <n>` / `--offset <n>` raw paging
 
-- `everr ci grep --pattern <text>` — search failing step logs for a literal, case-insensitive pattern across the last 7 days, excluding the current branch
-  - `--job-name <job>` / `--step-number <n>` target a specific step
-  - `--branch <name>` scope to a branch
-
 Collection-style commands accept `--limit <n>` and `--offset <n>` for pagination.
 
 ## Local Telemetry
